# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "药妆智多星多智能体客服系统",
    "version": "5.5"
  },
  "paths": {
...
    "/api/chat/image": {
      "post": {
        "summary": "Chat With Image",
        "description": "多模态图片分析接口",
        "operationId": "chat_with_image_api_chat_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_chat_with_image_api_chat_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/multimodal/stream": {
      "post": {
        "summary": "Stream Multimodal Chat",
        "description": "多模态图片分析 SSE 流式接口",
        "operationId": "stream_multimodal_chat_api_chat_multimodal_stream_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_stream_multimodal_chat_api_chat_multimodal_stream_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/voice": {
      "post": {
        "summary": "Chat With Voice",
        "description": "语音识别 + 文字对话",
        "operationId": "chat_with_voice_api_chat_voice_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_chat_with_voice_api_chat_voice_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chat/file": {
      "post": {
        "summary": "Chat With File",
        "description": "统一文件上传对话接口（图片/视频/PDF/DOCX/文本）",
        "operationId": "chat_with_file_api_chat_file_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_chat_with_file_api_chat_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_chat_with_file_api_chat_file_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "session_token": {
            "type": "string",
            "title": "Session Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_chat_with_file_api_chat_file_post"
      },
      "Body_chat_with_image_api_chat_image_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "session_token": {
            "type": "string",
            "title": "Session Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_chat_with_image_api_chat_image_post"
      },
      "Body_chat_with_voice_api_chat_voice_post": {
        "properties": {
          "audio": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Audio"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "session_token": {
            "type": "string",
            "title": "Session Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "audio"
        ],
        "title": "Body_chat_with_voice_api_chat_voice_post"
      },
      "Body_stream_multimodal_chat_api_chat_multimodal_stream_post": {
        "properties": {
          "image": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Image"
          },
          "query": {
            "type": "string",
            "title": "Query",
            "default": ""
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "default": ""
          },
          "session_token": {
            "type": "string",
            "title": "Session Token",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "image"
        ],
        "title": "Body_stream_multimodal_chat_api_chat_multimodal_stream_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}